+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/services/AnswerService.java
@@ -1,10 +1,13 @@
 package ru.dpopkov.knowthenics.services;
 
 import ru.dpopkov.knowthenics.model.Answer;
+import ru.dpopkov.knowthenics.model.KeyTerm;
 
 import java.util.Set;
 
 public interface AnswerService extends CrudService<Answer, Long> {
 
     Set<Answer> findAllByWordingEnLike(String searchString);
+
+    Set<Answer> findByKeyTerm(KeyTerm keyTerm);
 }