+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/services/QuestionService.java
@@ -1,5 +1,6 @@
 package ru.dpopkov.knowthenics.services;
 
+import ru.dpopkov.knowthenics.model.KeyTerm;
 import ru.dpopkov.knowthenics.model.Question;
 
 import java.util.Set;
@@ -7,4 +8,6 @@ import java.util.Set;
 public interface QuestionService extends CrudService<Question, Long> {
 
     Set<Question> findAllByWordingEnLike(String searchPattern);
+
+    Set<Question> findByKeyTerm(KeyTerm keyTerm);
 }