+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/services/map/AnswerMapService.java
@@ -3,6 +3,7 @@ package ru.dpopkov.knowthenics.services.map;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
 import ru.dpopkov.knowthenics.model.Answer;
+import ru.dpopkov.knowthenics.model.KeyTerm;
 import ru.dpopkov.knowthenics.services.AnswerService;
 
 import java.util.Set;
@@ -17,5 +18,11 @@ public class AnswerMapService extends AbstractMapService<Answer> implements Answ
         return super.findAll().stream()
                 .filter(a -> a.getWordingEn() != null && a.getWordingEn().contains(searchString))
                 .collect(Collectors.toSet());
+    }
+
+    @Override
+    public Set<Answer> findByKeyTerm(KeyTerm keyTerm) {
+        // todo: implement
+        throw new RuntimeException("Not implemented yet");
     }
 }